1.Attendance do not take a twice.(Attendance In hoi to Out Aavu jooye and Out hoi to in ) same date ma barabar che but hu login aaje kru chu ane logout kale kru chu not working properly 2.Holiday hoi to leave delete ane je divas ni  holiday hoi ae dvse leave accept na krva devu Joye. 3.Payment Appiye pachi delete krie ae to report wrong aave che.Worker as well as Employee. 4.Employee and worker na bne report match krva proper entery nakhi Pyament aapi Payment  delete kri pachi sacha aave che ke nai ae check krvu. 5.Rating aapiy ana 7 dvs pachi kai pn change na krva devu joye. 6.After Conversion leave add edit ane delete na krv devu joye as well as holiday ma pn.and as well as payment ma. 7.Attendance login na hoi to assign worker superwisor na kri ske ane checker ane payer attendace na lai ske. 8.worker ni evening attendance laye tyare.Monthly/Daily/unit and Hourly ma proper waya ma nai ane payment aapiye aetlu jma thavu joye.

+++ AttendanceSystem/Areas/Admin/Controllers/EmployeeRatingController.cs
@@ -13,7 +13,7 @@ namespace AttendanceSystem.Areas.Admin.Controllers
     public class EmployeeRatingController : Controller
     {
         AttendanceSystemEntities _db;
-         
+
         public EmployeeRatingController()
         {
             _db = new AttendanceSystemEntities();
@@ -137,6 +137,12 @@ namespace AttendanceSystem.Areas.Admin.Controllers
                         if (employeeRatingVM.EmployeeRatingId > 0)
                         {
                             tbl_EmployeeRating objEmployeeRating = _db.tbl_EmployeeRating.Where(x => x.EmployeeRatingId == employeeRatingVM.EmployeeRatingId).FirstOrDefault();
+                            if ((CommonMethod.CurrentIndianDateTime() - objEmployeeRating.CreatedDate).TotalDays > 7)
+                            {
+                                ModelState.AddModelError(" ", ErrorMessage.EmployeeRatingCanNotModifyAfter7Days);
+                                employeeRatingVM.EmployeeList = GetEmployeeList();
+                                return View(employeeRatingVM);
+                            }
                             objEmployeeRating.RateMonth = employeeRatingVM.RateMonth;
                             objEmployeeRating.RateYear = employeeRatingVM.RateYear;
                             objEmployeeRating.BehaviourRate = employeeRatingVM.BehaviourRate;
@@ -249,10 +255,18 @@ namespace AttendanceSystem.Areas.Admin.Controllers
                 }
                 else
                 {
-                    _db.tbl_EmployeeRating.Remove(objEmployeeRating);
-                    _db.SaveChanges();
+                    if ((CommonMethod.CurrentIndianDateTime() - objEmployeeRating.CreatedDate).TotalDays > 7)
+                    {
+                        ReturnMessage = "cannotmodify";
+                    }
+                    else
+                    {
 
-                    ReturnMessage = "success";
+                        _db.tbl_EmployeeRating.Remove(objEmployeeRating);
+                        _db.SaveChanges();
+
+                        ReturnMessage = "success";
+                    }
                 }
             }
             catch (Exception ex)
